Some platform support for libgit2

+++ build-libgit2.sh
@@ -4,5 +4,15 @@ mkdir -p libgit2/build
 cd libgit2/build
 cmake ..
 cmake --build .
-rm -f ../../project/addons/git_for_godot/gdnative/linuxbsd/libgit2.so*
-cp libgit2.so* ../../project/addons/git_for_godot/gdnative/linuxbsd/
+
+# Linux
+if [[ $(uname) == "Linux" ]]; then
+    rm -f ../../project/addons/git_for_godot/gdnative/linuxbsd/libgit2.so*
+    cp libgit2.so* ../../project/addons/git_for_godot/gdnative/linuxbsd/
+fi
+
+# macOS
+if [[ $(uname) == "Darwin" ]]; then
+    rm -f ../../project/addons/git_for_godot/gdnative/macos/libgit2*.dylib
+    cp libgit2*.dylib ../../project/addons/git_for_godot/gdnative/macos/
+fi
